Bugfix/21 dapper extension mapping (#76)
* Updated migrations.

* Re-worked IReadOnlyRepository interface. It was overprivaledged and the IQueryable interface needed to be promoted to the ILinqRepository interface. This will cause a breaking change but will be easily resolved by refactoring IReadOnlyRepository to ILinqRepository

* Revised persistence repository interfaces to use a more composite style.

* Refactorings to support better logging. Fixed bug in nested local events of business entities. Dapper integration tests nearly complete.

* Fixed logger issue with Ehab.

+++ Src/RCommon.Persistence.EfCore/EFCoreRepository.cs
@@ -135,12 +135,12 @@
             return queryable;
         }
 
-        public async override Task<int> GetCountAsync(ISpecification<TEntity> selectSpec, CancellationToken token = default)
+        public async override Task<long> GetCountAsync(ISpecification<TEntity> selectSpec, CancellationToken token = default)
         {
             return await this.FindCore(selectSpec.Predicate).CountAsync(token);
         }
 
-        public async override Task<int> GetCountAsync(Expression<Func<TEntity, bool>> expression, CancellationToken token = default)
+        public async override Task<long> GetCountAsync(Expression<Func<TEntity, bool>> expression, CancellationToken token = default)
         {
             return await this.FindCore(expression).CountAsync(token);
         }
